feat: create basic vision processing pipeline
This commit adds a new vision processing pipeline. This allows for the detection of objects and usage of depth estimation.

NOTE: The depth estimation still needs to be calibrated and some of the other features still need to be added, like distortion removal.

Additionally, support still needs to be added for camera calibration data to be used to estimate the position of the object along the X and Y axis.